Add new Build command; refactor tests
