+++ jib-cli/README.md
@@ -16,8 +16,9 @@ Use the `application` plugin's `installDist` task to create a runnable installat
 ```sh
 $ ../gradlew installDist
 $ ./build/install/jib/bin/jib
-Missing required parameters: base-image, destination-image
-Usage: jib (-d | -r) [-kv] [-c=<creationTime>] [-u=user] [-a=<arguments>[,
+Missing required subcommand
+Usage: jib [-kv] [-C=helper]... COMMAND
+A tool for creating container images.
 ...
 ```
 
@@ -28,7 +29,7 @@ Usage: jib (-d | -r) [-kv] [-c=<creationTime>] [-u=user] [-a=<arguments>[,
 The following example creates an nginx-based container to serve static content that is found in `path/to/website`.
 The result is loaded to the local Docker daemon as `my-static-website`:
 
-    $ ./build/install/jib/bin/jib \
+    $ ./build/install/jib/bin/jib build \
       --docker \
       nginx \
       my-static-website \
@@ -42,16 +43,15 @@ The result is loaded to the local Docker daemon as `my-static-website`:
 The following example uses _jib_ to containerize itself.  The image is pushed to a registry at `localhost:5000`:
 
     $ ../gradlew installDist
-    $ ./build/install/jib/bin/jib \
+    $ ./build/install/jib/bin/jib --insecure build \
       --registry \
       gcr.io/distroless/java \
       localhost:5000/jib:latest \
-      --insecure \
-      --entrypoint "java,-cp,/app/libs/*,com.google.cloud.tools.jib.cli.Cram" \
-      build/install/jib/lib:/app/libs
+      --entrypoint "java,-cp,/app/lib/*,com.google.cloud.tools.jib.cli.JibCli" \
+      build/install/jib/lib,/app/lib
     $ docker run --rm localhost:5000/cram:latest
 
 We need to use `--insecure` assuming the local registry does not support SSL.
 
-Note that we'd be better off using `jib-maven-plugin` to create the container since it would create better layer strategy that negates the need to use a fatjar.
-
+(Note that we'd be better off using `jib-gradle-plugin` to create the container since it would create better layer strategy that negates the
+need to use a fatjar.)
